woocommerce checkout page

+++ lib/common/utils.dart
@@ -3,13 +3,11 @@ import 'package:intl/intl.dart';
 class Utils {
   static String thousandSeperate(String numberString){
   try{
-  int number = int.parse(numberString);
+  double number = double.parse(numberString);
   return NumberFormat('#,###').format(number);
   }catch(ex){
     print(ex.toString() + "unable to parse number : "+ numberString);
     return "";
   }
-  
   }
- 
 }
